refactor(panel-widget): extract StepProgressView from CollectionLogHelperPanel shell
Moves the multi-step guidance progress label, required-items icon strip
(green/yellow/red availability borders), and Next Step / Skip buttons
into StepProgressView.  Shell delegates showStep(), hide(), and
setCallbacks() to it; scaleImage() helper is now private to the widget.

Part of A1b shell widget decomposition.
